First commit of the angle generator
A Makefile is provided for simulation and synthesis into an ICE40_hx4k.
TODO1: use an internal RAM rather than signal of array.
TODO2: pipe line the 64 bits (if reg_size=32) addition, in order to raise the frequency.
TODO3: write the real frequencies rather than test values, and include the master clock.
TODO4: improve the data monitor to compare to the frequencies.
TODO5: improve the data monitor to write a report on the derivate stability