Cordic conversion Y to 0
More code to run and bundle the convert-ion Y to 0.

The quick test is to connect the output of the Z to 0 bundle
to input of the Y to 0 without the pre-filter nor the down-sampling nor the filter.
It looks like working for most of the examples points using my wave viewer.
The 6 left hexa digits of the X converge to the same value (positive).
The 6 left hexa digits of the Y converge to 0 or FFFFFF.
The Z is a random value for a quick watch.