Cordic intermediary stages monitoring
New components to monitor individually the intermediary stages.
They compute the determinant of the vector (X and Y) before and after.
They compute the Z difference.
For a given stage, the absolute value of the results should be a constant
(Indeed, the chosen algorithm spins by +1 or -1, not +1, 0 or -1).

Left undone for now:
Find why the first stage is bad.
It is probably due to the interface between the angle gene and the stage
or a handling problem of "large" numbers.

Improve the coverage, as the number of full cycles should be the whose
make a full turn of a vector.

Check with large X and Y inputs to verify the limits.