Test strategy change for the RAM tied to 0 tests.
Due to the limited precision, the results are never exactly equal
  to what is expected.
The strategy was to check the ratio and quote the minimum and the maximum.
That makes false alarms with very low numbers.
Now the expected result is computed, and compared to the actual result.
The number of exact equalities, of plus one differences,
  minus one differences, and other is counted and displayed at the end.