Cordic Z to 0 first stage refactoring
There was a nasty bug.
The previous commit set a Q&D for the Z. This one has been fixed for X, Y and Z.

Now, the Z converge to 0, the reminder is due to the limited number of iterations.
Now, the module of X and Y is stable.

To spare 2 registers, the prepossess was computed in parallel mode.
In fact it spared only 1 register as the input Y is intended for tests only.
In run mode, it is loaded with 0's and the 0's are shifted. Then the register is simplified.

Each stage loads the (next) behavior when reg_sync is high, and run otherwise.
Since the first stage set its output registers during the reg_sync, the first intermediary stage latched the behavior on the previous sample.

Another fix is the intermediary stages should start at 2, means the angle is about 22 degrees.
Since the angle is moved to [ -PI/4 PI/4 ], tg 1 (PI/4) is irrelevant.