End to end test
It is a pretty good verification of the Cordic elements.
* The first set converges the angle to 0.
The amplitude of the vector grows by 16.44% regardless everything.
* The second set returns the vector to the horizontal axis, positive side.
In the future, a descrambling is going to show the angle spins regularly.
At each result, the angles of the input test cases at 0, 90, 180 and 240
  are separated by $40..0 to each other.
The amplitude of the output vector (the X value) grows by 35.59%
  regardless everything.