End to end DC test automated test
There are four instantiations with a starting vector at 0, 90, 180 and 240 degrees.
The automated test checks the module of the output vectors, for all the instances and for all the simulation grows by a calculated constant.
The values are about 16.44% for the first set Z to 0 and 35.59% for both sets Z to 0 followed by Y to 0.

The automated test checks the fours instances, for each time should have a 90 degrees difference to each other.