Cordic end to end test mostly finished
The chain, without the filtering nor the down-sampling is close to be ready for service.
The vectors spin properly and spin back to the X axis.
At the end, the Z, for each combination note octave, grows continuously
  with a coherent delta. The delta looks like to be coherent between
  combinations note octaves as well.
The meta data, however, comes too early.
An automated check still has to be written.
The improvement of the angle generator has to be done.